cmd/oscap-customizations: add initial helpers
Add some helper methods to generate OpenSCAP customizations on the fly.
This function accepts needs the profile name, distro and an optional
`json` tailoring string. If the tailoring string is empty, the functions
will generate the standard customization options for the given profile.
If the tailoring contents are not empty, the returned customizations
should exclude the options specified in the tailoring file.

For example:
If `partition_for_var` is provided as an excluded tailoring option the
generated filesystem customizations should also exclude the `var`
partition.